+++ crates/base/src/global/veci8_l2.rs
@@ -61,8 +61,8 @@ impl GlobalProductQuantization for Veci8L2 {
     type ProductQuantizationL2 = Veci8L2;
 
     fn product_quantization_distance(
-        _dims: u16,
-        _ratio: u16,
+        _dims: u32,
+        _ratio: u32,
         _centroids: &[Scalar<Self>],
         _lhs: Borrowed<'_, Self>,
         _rhs: &[u8],
@@ -70,8 +70,8 @@ impl GlobalProductQuantization for Veci8L2 {
         unimplemented!()
     }
     fn product_quantization_distance2(
-        _dims: u16,
-        _ratio: u16,
+        _dims: u32,
+        _ratio: u32,
         _centroids: &[Scalar<Self>],
         _lhs: &[u8],
         _rhs: &[u8],
@@ -79,8 +79,8 @@ impl GlobalProductQuantization for Veci8L2 {
         unimplemented!()
     }
     fn product_quantization_distance_with_delta(
-        _dims: u16,
-        _ratio: u16,
+        _dims: u32,
+        _ratio: u32,
         _centroids: &[Scalar<Self>],
         _lhs: Borrowed<'_, Self>,
         _rhs: &[u8],
